# docs/auth_service_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Saturn MouseHunter Auth Service",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/roles/{role_id}/permissions": {
      "get": {
        "tags": [
          "角色管理"
        ],
        "summary": "Get Role With Permissions",
        "description": "获取角色及其权限",
        "operationId": "get_role_with_permissions_api_v1_admin_roles__role_id__permissions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Role Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleWithPermissions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/permissions/": {
      "post": {
        "tags": [
          "权限管理"
        ],
        "summary": "Create Permission",
        "description": "创建权限",
        "operationId": "create_permission_api_v1_admin_permissions__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "权限管理"
        ],
        "summary": "List Permissions",
        "description": "获取权限列表",
        "operationId": "list_permissions_api_v1_admin_permissions__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "permission_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Permission Name"
            }
          },
          {
            "name": "resource",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "name": "is_system_permission",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is System Permission"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PermissionOut"
                  },
                  "title": "Response List Permissions Api V1 Admin Permissions  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/permissions/count": {
      "get": {
        "tags": [
          "权限管理"
        ],
        "summary": "Count Permissions",
        "description": "获取权限总数",
        "operationId": "count_permissions_api_v1_admin_permissions_count_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "permission_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Permission Name"
            }
          },
          {
            "name": "resource",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "name": "is_system_permission",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is System Permission"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Count Permissions Api V1 Admin Permissions Count Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/permissions/system": {
      "get": {
        "tags": [
          "权限管理"
        ],
        "summary": "List System Permissions",
        "description": "获取系统权限列表",
        "operationId": "list_system_permissions_api_v1_admin_permissions_system_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PermissionOut"
                  },
                  "type": "array",
                  "title": "Response List System Permissions Api V1 Admin Permissions System Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/permissions/by-resource/{resource}": {
      "get": {
        "tags": [
          "权限管理"
        ],
        "summary": "List Permissions By Resource",
        "description": "根据资源获取权限列表",
        "operationId": "list_permissions_by_resource_api_v1_admin_permissions_by_resource__resource__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "resource",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PermissionOut"
                  },
                  "title": "Response List Permissions By Resource Api V1 Admin Permissions By Resource  Resource  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/permissions/check/{permission_code}": {
      "get": {
        "tags": [
          "权限管理"
        ],
        "summary": "Check Permission Exists",
        "description": "检查权限是否存在",
        "operationId": "check_permission_exists_api_v1_admin_permissions_check__permission_code__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "permission_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Permission Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Check Permission Exists Api V1 Admin Permissions Check  Permission Code  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/permissions/{permission_id}": {
      "get": {
        "tags": [
          "权限管理"
        ],
        "summary": "Get Permission",
        "description": "获取权限详情",
        "operationId": "get_permission_api_v1_admin_permissions__permission_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "permission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Permission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "权限管理"
        ],
        "summary": "Update Permission",
        "description": "更新权限",
        "operationId": "update_permission_api_v1_admin_permissions__permission_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "permission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Permission Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "权限管理"
        ],
        "summary": "Delete Permission",
        "description": "删除权限",
        "operationId": "delete_permission_api_v1_admin_permissions__permission_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "permission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Permission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PermissionIn": {
        "properties": {
          "permission_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Permission Name",
            "description": "权限名称"
          },
          "permission_code": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Permission Code",
            "description": "权限编码"
          },
          "resource": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Resource",
            "description": "资源名称"
          },
          "action": {
            "type": "string",
            "maxLength": 50,
            "minLength": 2,
            "title": "Action",
            "description": "操作类型"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "权限描述"
          },
          "is_system_permission": {
            "type": "boolean",
            "title": "Is System Permission",
            "description": "是否系统权限",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "permission_name",
          "permission_code",
          "resource",
          "action"
        ],
        "title": "PermissionIn",
        "description": "权限创建模型"
      },
      "PermissionOut": {
        "properties": {
          "permission_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Permission Name",
            "description": "权限名称"
          },
          "permission_code": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Permission Code",
            "description": "权限编码"
          },
          "resource": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Resource",
            "description": "资源名称"
          },
          "action": {
            "type": "string",
            "maxLength": 50,
            "minLength": 2,
            "title": "Action",
            "description": "操作类型"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "权限描述"
          },
          "is_system_permission": {
            "type": "boolean",
            "title": "Is System Permission",
            "description": "是否系统权限",
            "default": false
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "权限ID"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新时间"
          }
        },
        "type": "object",
        "required": [
          "permission_name",
          "permission_code",
          "resource",
          "action",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "PermissionOut",
        "description": "权限输出模型"
      },
      "PermissionUpdate": {
        "properties": {
          "permission_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Permission Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "PermissionUpdate",
        "description": "权限更新模型"
      },
...
      "RoleScope": {
        "type": "string",
        "enum": [
          "GLOBAL",
          "TENANT",
          "SYSTEM"
        ],
        "title": "RoleScope",
        "description": "角色范围"
      },
...
      "RoleWithPermissions": {
        "properties": {
          "role_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Role Name",
            "description": "角色名称"
          },
          "role_code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 2,
            "title": "Role Code",
            "description": "角色编码"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "角色描述"
          },
          "scope": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RoleScope"
              }
            ],
            "description": "角色范围",
            "default": "GLOBAL"
          },
          "is_system_role": {
            "type": "boolean",
            "title": "Is System Role",
            "description": "是否系统角色",
            "default": false
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "是否激活",
            "default": true
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "角色ID"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新时间"
          },
          "permissions": {
            "items": {
              "$ref": "#/components/schemas/PermissionOut"
            },
            "type": "array",
            "title": "Permissions",
            "description": "角色权限列表"
          }
        },
        "type": "object",
        "required": [
          "role_name",
          "role_code",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "RoleWithPermissions",
        "description": "包含权限的角色模型"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}